+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Add Block: A version container keeping various Concepts, Pairs and Data 
 - Add Data: A value set to a specific Pair 
 - Add Pair: Relationship between two Concepts
 - Add Concept:  Smallest unit of semantics